move DictArray and related functionality from vortex-dict to vortex-array
vortex-dict reexports the vortex-array functionality, so there should be no
user-facing changes. This code movement is necessary so that vortex-array can
convert from dictionary-encoded arrow arrays to vortex DictArrays. vortex-array
cannot depend on vortex-dict since this introduces a circular dependency.
