+++ vortex-array/goldenfiles/dict.metadata
