+++ vortex-array/src/arrays/dict/compute/is_constant.rs
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-use vortex_array::compute::{
+use crate::compute::{
     IsConstantKernel, IsConstantKernelAdapter, IsConstantOpts, is_constant_opts,
 };
-use vortex_array::register_kernel;
+use crate::register_kernel;
 use vortex_error::VortexResult;
 
-use crate::{DictArray, DictVTable};
+use crate::arrays::{DictArray, DictVTable};
 
 impl IsConstantKernel for DictVTable {
     fn is_constant(&self, array: &DictArray, opts: &IsConstantOpts) -> VortexResult<Option<bool>> {
