+++ vortex-array/src/arrays/dict/compute/is_sorted.rs
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-use vortex_array::compute::{IsSortedKernel, IsSortedKernelAdapter, is_sorted, is_strict_sorted};
-use vortex_array::register_kernel;
+use crate::compute::{IsSortedKernel, IsSortedKernelAdapter, is_sorted, is_strict_sorted};
+use crate::register_kernel;
 use vortex_error::VortexResult;
 
-use crate::{DictArray, DictVTable};
+use crate::arrays::{DictArray, DictVTable};
 
 impl IsSortedKernel for DictVTable {
     fn is_sorted(&self, array: &DictArray) -> VortexResult<bool> {
