+++ vortex-array/src/arrays/mod.rs
@@ -10,6 +10,7 @@ mod bool;
 mod chunked;
 mod constant;
 mod datetime;
+pub mod dict;
 mod extension;
 mod list;
 mod null;
@@ -29,6 +30,7 @@ pub use self::chunked::*;
 pub use self::constant::*;
 pub use self::datetime::*;
 pub use self::decimal::*;
+pub use self::dict::*;
 pub use self::extension::*;
 pub use self::list::*;
 pub use self::null::*;
